rabbit_plugins: Filter non-plugins out
When the module discovers plugins in the plugins directories, it
considers all applications there. There are actual RabbitMQ plugins,
their dependencies, and possible other applications which are neither
plugins nor dependencies.

So far, this was the case of `syslog` which could be used by Lager if
the user configures it. To make sure `rabbit_plugins` or
rabbitmq-plugins(8) didn't mess with `syslog`, there was a hard-coded
filtering function which excluded this specific application.

In RabbitMQ Enterprise Edition, we want to add another applicatino of
this kind: inet_tcp_compress_dist. This application implements an Erlang
distribution protocol, so it is effectively unrelated to RabbitMQ.
However, to keep packaging simple, it is provided in the plugins
directory.

We could add it to the hard-coded filtering function, but instead this
function is replaced by the following logic to filter plugins:
  * Applications which depend on `rabbit` are considered actual plugins.
  * Plugins' dependencies, direct and indirect, are considered plugins as well.

Therefore, `rabbit_plugins` will ignore:
  * Applications which don't depend on `rabbit` and no plugins
    depend on them (again, directly or indirectly)
  * Applications which are part of Erlang/OTP.

The code will log the following debug messages to list ignored
applications:

2020-01-14 13:18:41.468 [debug] <0.941.0> Plugins discovery: ignoring getopt, not a RabbitMQ plugin
2020-01-14 13:18:41.468 [debug] <0.941.0> Plugins discovery: ignoring inet_tcp_compress_dist, not a RabbitMQ plugin
2020-01-14 13:18:41.468 [debug] <0.941.0> Plugins discovery: ignoring lz4, not a RabbitMQ plugin
2020-01-14 13:18:41.468 [debug] <0.941.0> Plugins discovery: ignoring syslog, not a RabbitMQ plugin
2020-01-14 13:18:41.468 [debug] <0.941.0> Plugins discovery: ignoring zstd, not a RabbitMQ plugin